fix: corrige erros e implementa metodo de exclusao de consultas corretamente

+++ src/main/java/com/senai/clinicaApi/services/ConsultaService.java
@@ -104,6 +104,9 @@ public class ConsultaService {
     }
 
     public boolean excluirConsulta(long id) {
+        if (!consultaRepository.existsById(id)) {
+            return false;
+        }
         consultaRepository.deleteById(id);
         return true;
     }
